♻️ refactor: App router to Router.jsx
